+++ packages/plugin-highlight/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change log
 
+## bytemd-plugin-highlight@0.0.4
+
+> 2026-07-30
+
+- This package no longer provides `@tofrankie/bytemd-plugin-highlight/scss/<theme>` exports
+- Theme CSS is now generated from `@primer/primitives` and published from `dist/styles`
+- Package style exports still use `@tofrankie/bytemd-plugin-highlight/styles/<theme>.css`
+
 ## bytemd-plugin-highlight@0.0.3
 
 > 2026-07-28